`Sign Up` feature
Fix auth modal so it populates with `sign up` and `sign in` forms on
click of `log In` button.

Create: api.js
        events.js
        ui.js
    (add sign up functions to all of the above)

Modify: config.js - add localhost:4741 to later host api locally
        index.js - add events.addHandlers() to allow click events to run

Sign Up is not yet functional, as API needs to be built and connected
Need to add message divs to the auth modal
      - sign up error
      - sign in error (will hopefully be auto sign in)
      - eventually: password and sign out errors

Also need to finish auto sign-in feature along with sign in.