add js function that adds product name and price to cart
function is triggered by a product's add-to-cart button. Using jQuery, takes the event.target and traverse up the DOM till it finds .product class, then fetches the value of .prod-name/.product-price

Values are stored in local name and price variables respectively. then appended in a new table row in .fill-this

coded with @lindsayka9 and @axb6452