Modify purchase logic and move checkout function UI modifications to UI
Changes the stripe token callback function to send a patch request to carts resource, then sends the post request to /charge. This ensures that the order is only updated when a user has finished entering their details.

Also moved hiding the update and delete cart item button to ui.checkoutSuccess so that a failed checkout(if that ever happens) doesn't hide any buttons.